add extract rule and update dag
