+++ autoafids/workflow/rules/cnn.smk
@@ -9,6 +9,9 @@ def get_model():
 
     return (Path(download_dir)/ "model" / Path(local_model).name).absolute()
 
+def get_extracted_models():
+    return (Path(download_dir)/ "extracted_model/").absolute()
+
 rule download_cnn_model:
     params:
         url=config["resource_urls"][config["model"]],
@@ -18,6 +21,14 @@ rule download_cnn_model:
     shell:
         "mkdir -p {params.model_dir} && wget https://{params.url} -O {output}"
 
+rule extract_models:
+    input:
+        model_path = get_model()
+    output:
+        model_dir = directory(get_extracted_models())
+    script:
+        "../scripts/extract_models.py"
+    
 rule gen_fcsv:
     input:
         t1w = bids(
@@ -27,8 +38,8 @@ rule gen_fcsv:
                 res=config["res"],
                 suffix="T1w.nii.gz",
                 **inputs["t1w"].wildcards,
-                ), 
-        model = get_model(),
+                ),
+        extracted_model = get_extracted_models(),
         prior = bids(
             root=work,
             datatype="registration",
@@ -37,6 +48,8 @@ rule gen_fcsv:
             suffix="afids.fcsv",
             **inputs["t1w"].wildcards,
         ),
+    params:
+        model_path = get_model()
     output:
         fcsv = bids(
             root=root,
@@ -52,4 +65,4 @@ rule gen_fcsv:
             **inputs["t1w"].wildcards
         ),
     shell:
-        'auto_afids_cnn_apply {input.t1w} {input.model} {output.fcsv} {input.prior}'
+        'auto_afids_cnn_apply {input.t1w} {params.model_path} {input.extracted_model} {output.fcsv} {input.prior}'
